Fix 14d and 28d prediction accuracy: remove noisy T7 features from 14d, add NaN imputation for 28d
14d model: T7 geopolitical features added noise at this horizon (MAPE 0.149→0.138,
R² 0.672→0.700 without them). Reverted to T1+T2+T4 feature set.

28d model: 7 features had >10% NaN (enso_lag12m 67%, rain_anomaly 33%), causing
dropna to discard 2/3 of training rows (546→177) and yielding only 5 CV folds
with R²=-3.45. Added median imputation and lowered min_train from 104 to 60,
giving 37 folds and R²=+0.64.

Also fixed trade data period parsing for float-string values like "201401.0".

+++ pipeline/assemble.py
@@ -71,7 +71,7 @@ def build_daily_df() -> pd.DataFrame:
     guatemala = read_table("trade_guatemala", conn)
     if not guatemala.empty:
         guatemala["date"] = pd.to_datetime(guatemala["period"].apply(
-            lambda p: pd.Timestamp(year=int(str(int(p))[:4]), month=int(str(int(p))[4:6]), day=1)
+            lambda p: pd.Timestamp(year=int(str(float(p))[:4]), month=int(str(float(p))[4:6]), day=1)
             + pd.offsets.MonthEnd(0)
         ))
         gt = guatemala[["date", "qty_kg", "net_wgt_kg", "value_usd"]].copy()
@@ -86,7 +86,7 @@ def build_daily_df() -> pd.DataFrame:
     saudi = read_table("trade_saudi", conn)
     if not saudi.empty:
         saudi["date"] = pd.to_datetime(saudi["period"].apply(
-            lambda p: pd.Timestamp(year=int(str(int(p))[:4]), month=int(str(int(p))[4:6]), day=1)
+            lambda p: pd.Timestamp(year=int(str(float(p))[:4]), month=int(str(float(p))[4:6]), day=1)
             + pd.offsets.MonthEnd(0)
         ))
         sa = saudi[["date", "net_wgt_kg", "value_usd"]].copy()
